add rollup to support multiple files
