+++ package.json
@@ -11,7 +11,7 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
-    "build": "sh scripts/build.sh",
+    "build": "rollup -c",
     "test": "jest --runInBand",
     "test:updateSnapshots": "jest --runInBand -u"
   },
@@ -26,6 +26,7 @@
   },
   "homepage": "https://github.com/rune/rune-games-sdk#readme",
   "devDependencies": {
+    "@rollup/plugin-typescript": "^8.3.1",
     "@types/jest": "^27.0.2",
     "@types/node": "^16.9.6",
     "@typescript-eslint/eslint-plugin": "^5.13.0",
@@ -34,6 +35,7 @@
     "eslint-config-prettier": "^8.1.0",
     "jest": "^26.6.3",
     "prettier": "^2.2.1",
+    "rollup": "^2.70.1",
     "ts-jest": "^26.5.4",
     "typescript": "^4.5.5"
   },
