+++ src/browser.ts
@@ -1,0 +1,2 @@
+import {Rune} from './index';
+export default Rune;
